# backend/src/main/java/com/tradingsim/application/dto/GameDecisionRequest.java
package com.tradingsim.application.dto;

import java.math.BigDecimal;

/**
 * 游戏决策请求DTO
 * 
 * @author TradingSim Team
 */
public class GameDecisionRequest {
    
    private String sessionId;
    private Integer frameIndex;
    private String decisionType;
    private BigDecimal price;
    private Integer quantity;
    private Long responseTimeMs;

    public GameDecisionRequest() {}

    public GameDecisionRequest(String sessionId,Integer frameIndex, String decisionType, BigDecimal price, Integer quantity, Long responseTimeMs) {
        this.sessionId = sessionId;
        this.frameIndex = frameIndex;
        this.decisionType = decisionType;
        this.price = price;
        this.quantity = quantity;
        this.responseTimeMs = responseTimeMs;
    }

    // Getters and Setters
    public String getSessionId() {
        return sessionId;
    }

    public void setSessionId(String sessionId) {
        this.sessionId = sessionId;
    }
    public Integer getFrameIndex() { return frameIndex;}
    public void setFrameIndex(Integer frameIndex) { this.frameIndex = frameIndex;}
    public String getDecisionType() {
        return decisionType;
    }

    public void setDecisionType(String decisionType) {
        this.decisionType = decisionType;
    }

    public BigDecimal getPrice() {
        return price;
    }

    public void setPrice(BigDecimal price) {
        this.price = price;
    }

    public Integer getQuantity() {
        return quantity;
    }

    public void setQuantity(Integer quantity) {
        this.quantity = quantity;
    }

    public Long getResponseTimeMs() {
        return responseTimeMs;
    }

    public void setResponseTimeMs(Long responseTimeMs) {
        this.responseTimeMs = responseTimeMs;
    }
}