# docs/api-docs.json
{
"openapi": "3.0.1",
"info": {
"title": "TradingSim API",
"description": "交易模拟系统后端API文档",
"contact": {
"name": "TradingSim Team",
"url": "https://github.com/tradingsim/tradingsim",
"email": "support@tradingsim.com"
},
"license": {
"name": "MIT License",
"url": "https://opensource.org/licenses/MIT"
},
"version": "1.0.0"
},
"servers": [
{
"url": "http://localhost:8080",
"description": "开发环境"
},
{
"url": "https://api.tradingsim.com",
"description": "生产环境"
}
],
"security": [
{
"Bearer Authentication": []
}
],
...
"paths": {
...
"/game/sessions": {
"post": {
"tags": [
"游戏管理"
],
"summary": "创建游戏会话",
"description": "创建新的交易模拟游戏会话",
"operationId": "createSession",
"parameters": [
{
"name": "stockCode",
"in": "query",
"description": "股票代码",
"required": true,
"schema": {
"type": "string"
},
"example": "AAPL"
},
{
"name": "difficulty",
"in": "query",
"description": "游戏难度",
"required": true,
"schema": {
"type": "string"
},
"example": "EASY"
}
],
"responses": {
"200": {
"description": "游戏会话创建成功",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
},
"400": {
"description": "请求参数错误",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
}
}
}
},
"/game/sessions/{sessionId}/start": {
"post": {
"tags": [
"游戏管理"
],
"operationId": "startSession",
"parameters": [
{
"name": "sessionId",
"in": "path",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
}
}
}
},
"/game/sessions/{sessionId}/finish": {
"post": {
"tags": [
"游戏管理"
],
"operationId": "finishSession",
"parameters": [
{
"name": "sessionId",
"in": "path",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
}
}
}
},
"/game/sessions/{sessionId}/decisions": {
"get": {
"tags": [
"游戏管理"
],
"operationId": "getSessionDecisions",
"parameters": [
{
"name": "sessionId",
"in": "path",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"type": "array",
"items": {
"$ref": "#/components/schemas/GameDecisionResponse"
}
}
}
}
}
}
},
"post": {
"tags": [
"游戏管理"
],
"operationId": "submitDecision",
"parameters": [
{
"name": "sessionId",
"in": "path",
"required": true,
"schema": {
"type": "string"
}
}
],
"requestBody": {
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/GameDecisionRequest"
}
}
},
"required": true
},
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/GameDecisionResponse"
}
}
}
}
}
}
},
...
"/users/{userId}/game-history": {
"get": {
"tags": [
"用户管理"
],
"summary": "获取用户游戏历史",
"description": "获取指定用户的游戏历史记录",
"operationId": "getUserGameHistory",
"parameters": [
{
"name": "userId",
"in": "path",
"description": "用户ID",
"required": true,
"schema": {
"type": "string"
}
},
{
"name": "limit",
"in": "query",
"description": "返回记录数量限制",
"required": false,
"schema": {
"type": "integer",
"format": "int32",
"default": 10
},
"example": 10
}
],
"responses": {
"200": {
"description": "成功获取游戏历史",
"content": {
"*/*": {
"schema": {
"type": "array",
"items": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
}
},
"404": {
"description": "用户不存在",
"content": {
"*/*": {
"schema": {
"type": "array",
"items": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
}
}
}
}
},
...
"/game/users/{userId}/history": {
"get": {
"tags": [
"游戏管理"
],
"operationId": "getUserGameHistory_1",
"parameters": [
{
"name": "userId",
"in": "path",
"required": true,
"schema": {
"type": "string"
}
},
{
"name": "limit",
"in": "query",
"required": false,
"schema": {
"type": "integer",
"format": "int32",
"default": 20
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"type": "array",
"items": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
}
}
}
}
},
"/game/sessions/{sessionId}": {
"get": {
"tags": [
"游戏管理"
],
"operationId": "getSession",
"parameters": [
{
"name": "sessionId",
"in": "path",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
}
}
}
},
"/game/leaderboard": {
"get": {
"tags": [
"游戏管理"
],
"operationId": "getLeaderboard",
"parameters": [
{
"name": "limit",
"in": "query",
"required": false,
"schema": {
"type": "integer",
"format": "int32",
"default": 10
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"type": "array",
"items": {
"$ref": "#/components/schemas/GameSessionResponse"
}
}
}
}
}
}
}
},
...
},
"components": {
"schemas": {
...
"GameSessionResponse": {
"type": "object",
"properties": {
"sessionId": {
"type": "string"
},
"userId": {
"type": "string"
},
"stockCode": {
"type": "string"
},
"status": {
"type": "string"
},
"difficulty": {
"type": "string"
},
"initialBalance": {
"type": "number"
},
"currentBalance": {
"type": "number"
},
"totalReturn": {
"type": "number"
},
"returnPercentage": {
"type": "number"
},
"currentFrame": {
"type": "integer",
"format": "int32"
},
"totalFrames": {
"type": "integer",
"format": "int32"
},
"score": {
"type": "integer",
"format": "int32"
},
"createdAt": {
"type": "string",
"format": "date-time"
},
"updatedAt": {
"type": "string",
"format": "date-time"
},
"finishedAt": {
"type": "string",
"format": "date-time"
}
}
},
"GameDecisionRequest": {
"type": "object",
"properties": {
"sessionId": {
"type": "string"
},
"frameIndex": {
"type": "integer",
"format": "int32"
},
"decisionType": {
"type": "string"
},
"price": {
"type": "number"
},
"quantity": {
"type": "integer",
"format": "int32"
},
"responseTimeMs": {
"type": "integer",
"format": "int64"
}
}
},
"GameDecisionResponse": {
"type": "object",
"properties": {
"decisionId": {
"type": "string"
},
"sessionId": {
"type": "string"
},
"frameIndex": {
"type": "integer",
"format": "int32"
},
"decisionType": {
"type": "string"
},
"price": {
"type": "number"
},
"quantity": {
"type": "integer",
"format": "int32"
},
"responseTimeMs": {
"type": "integer",
"format": "int64"
},
"profit": {
"type": "number"
},
"profitPercentage": {
"type": "number"
},
"createdAt": {
"type": "string",
"format": "date-time"
}
}
},
...
},
"securitySchemes": {
"Bearer Authentication": {
"type": "http",
"description": "请在请求头中添加 Authorization: Bearer {token}",
"scheme": "bearer",
"bearerFormat": "JWT"
}
}
}
}